前缀和T7.和谐合影（三）
题目：
Description
FJ's N cows (2 <= N <= 100,000) are standing at various positions along a long one-dimensional fence. The ith cow is standing at position x_i (an integer in the range 0...1,000,000,000) and is either a plain white cow or a spotted cow. No two cows occupy the same position, and there is at least one white cow. FJ wants to take a photo of a contiguous interval of cows for the county fair, but in fairness to his different cows, he wants to ensure there are equal numbers of white and spotted cows in the photo. FJ wants to determine the maximum size of such a fair photo, where the size of a photo is the difference between the maximum and minimum positions of the cows in the photo. To give himself an even better chance of taking a larger photo, FJ has with him a bucket of paint that he can use to paint spots on an arbitrary subset of his white cows of his choosing, effectively turning them into spotted cows. Please determine the largest size of a fair photo FJ can take, given that FJ has the option of painting some of his white cows (of course, he does not need to paint any of the white cows if he decides this is better).
可以先任意把0染成1.
区间长度定义为,[L,R]中最右和最左的数的差的绝对值.
求一个最长区间,满足区间中所有数0和1的个数相同.
Input
 * Line 1: The integer N.
* Lines 2..1+N: Line i+1 contains x_i and either W (for a white cow) or S (for a spotted cow). 
Output
* Line 1: The maximum size of a fair photo FJ can take, after possibly painting some of his white cows to make them spotted.
Sample Input
5
8 W
11 S
3 W
10 W
5 S
INPUT DETAILS: There are 5 cows. One of them is a white cow at position 8, and so on.
Sample Output
7
OUTPUT DETAILS: FJ takes a photo of the cows from positions 3 to positions 10. 
There are 4 cows in this range -- 3 white and 1 spotted -- so he needs to paint one
of the white cows to make it spotted. 

思路（针对代码1）：
由于题目中只给了0和1，所以我们不方便比较，将0更换成-1，使得-1+1=0，方便前缀和处理
在初始化之后，我用q[tot]来记录前缀和为tot的起始位置。
其实感觉自己写的太麻烦了，不过可以参考参考代码2，提升一下思维空间。
因为题目中说过，我们可以将0变成1，那这里面就是可以将-1变成1，也就是q[i]和q[i+2]
所以我们要通过一遍循环来刷新一下q数组里边的值。
q[i]=max(q[i],q[i+2]);
巧妙地解决了刷漆的问题了。
最后的方法也就是再来最后一遍循环，根据q[i]来确定最大值。
输出即可。

代码2我来简单的讲一讲。
不看思路，可能很难看懂（我也钻研了将近1天）
适用于C++，将负数空间成功转移到了正数的空间。
可能看上去有些别扭，但是思路与代码1还是一样的。
同样也是不断更新q[i],q[i+2]的值。
q[tot]现在记得不是哪个点，而是在数轴上的距离。
tot之所以从n开始，是因为只有这样做，才能成功地把负数空间转移到正数空间。
我们可以看到在程序中，无论是建立还是查询，tot都是从n开始，这也验证了我的想法。
在一些特殊处理后，就出来了。
最后只要比较a[i]-q[tot]的值。

总结：在这道题中，我们见识了两种巧妙解题的方法。从代码1到代码2的过渡并不是突如其来，十分突兀的，而是有一种合理的转接过程。
今天的前缀和具有稍微提高性质的几道题也让我明白了如下几点：
1.总是粗心，补丁忘写，导致不能拿全分；
2.数据范围的定义必须合理准确，不多不少是最好，不仅是数组范围，而且也要考虑是否用int64，qword等；
3.理解题目总是按一种循序渐进的顺序，思维的条约要合理，易于思考，这样才能体会到算法的灵魂；
4.巧妙的算法要多看（如代码2，负-》正，一句tot=n即可，要是我肯定大动干戈了），多看才能自己写出来；
5.要善用学过的知识，考虑到题目中的任何边缘数据（俗称打补丁），使得程序往满分方向靠拢。